requirements/pypi.txt
django-taggit
djangorestframework
djangorestframework-filters
django-filter
